+++ CHANGELOG.md
@@ -30,7 +30,9 @@ audio focus was lost
 
 #### What's Changed
 - Ignore MediaStore tags is now on by default
-- Removed the "Play from genre" option in the library/detail playback mode settings
+- Removed the "Play from genre" option in the library/detail playback mode settings+
+- "Use alternate notification action" is now "Custom notification action"
+- "Show covers" and "Ignore MediaStore covers" have been unified into "Album covers"
 
 #### Dev/Meta
 - Completed migration to reactive playback system